requirements.txt
torch==1.6.0
numpy==1.19.2
seaborn==0.11.0
wandb==0.10.15
pandas==1.1.3
dl_thornet==0.0.1
matplotlib==3.3.2
psutil==5.7.2
FlowIO==0.9.8
tqdm==4.50.2
parc==0.31
PyYAML==5.4.1
rich==9.12.4
scikit_learn==0.24.1
